+++ MultiQuestensions/MultiQuestensions.vcxproj
@@ -22,6 +22,7 @@
   <ItemGroup>
     <ClCompile Include="src\Beatmaps\PreviewBeatmapPacket.cpp" />
     <ClCompile Include="src\Beatmaps\PreviewBeatmapStub.cpp" />
+    <ClCompile Include="src\Extensions\ExtendedPlayer.cpp.dis" />
     <ClCompile Include="src\main.cpp" />
     <ClCompile Include="src\Packets\PacketManager.cpp" />
     <ClCompile Include="src\Packets\PacketSerializer.cpp" />
@@ -52,11 +53,14 @@
     <ClInclude Include="extern\songloader\shared\API.hpp" />
     <ClInclude Include="include\Beatmaps\PreviewBeatmapPacket.hpp" />
     <ClInclude Include="include\Beatmaps\PreviewBeatmapStub.hpp" />
+    <ClInclude Include="include\Extensions\ExtendedPlayer.hpp.dis" />
+    <ClInclude Include="include\Hooks\Hooks.hpp" />
     <ClInclude Include="include\main.hpp" />
     <ClInclude Include="include\Packets\Callback.hpp" />
     <ClInclude Include="include\Packets\PacketManager.hpp" />
     <ClInclude Include="include\Packets\PacketSerializer.hpp" />
     <ClInclude Include="include\UI\LobbySetupPanel.hpp" />
+    <ClInclude Include="src\Hooks\NetworkPlayerEntitlementChecker.cpp" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
     <ProjectGuid>{1c603531-cd52-4be0-a840-ba695bb7e0c7}</ProjectGuid>
